requirements.txt
pillow~=9.3.0
numpy~=1.23.5
aiogram~=2.24
onnxruntime~=1.12.1